revno: 8809 message:   Changed MapView.opacity from int to double
committer: Lars Helge Øverland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Thu 2012-11-01 08:33:34 +0300

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/mapping/MapView.java
@@ -108,7 +108,7 @@ public class MapView
 
     private Integer radiusHigh;
 
-    private Integer opacity;
+    private Double opacity;
     
     private OrganisationUnitGroupSet organisationUnitGroupSet;
 
@@ -124,7 +124,7 @@ public class MapView
                     DataElementGroup dataElementGroup, DataElement dataElement,
                     Period period, OrganisationUnit parentOrganisationUnit, OrganisationUnitLevel organisationUnitLevel,
                     String legendType, Integer method, Integer classes, String colorLow, String colorHigh,
-                    MapLegendSet legendSet, Integer radiusLow, Integer radiusHigh, int opacity )
+                    MapLegendSet legendSet, Integer radiusLow, Integer radiusHigh, Double opacity )
     {
         this.layer = layer;
         this.name = name;
@@ -399,12 +399,12 @@ public class MapView
     @JsonProperty
     @JsonView( {DetailedView.class, ExportView.class} )
     @JacksonXmlProperty( namespace = Dxf2Namespace.NAMESPACE )
-    public Integer getOpacity()
+    public Double getOpacity()
     {
         return opacity;
     }
 
-    public void setOpacity( Integer opacity )
+    public void setOpacity( Double opacity )
     {
         this.opacity = opacity;
     }
